+++ src/app/layout.tsx
@@ -7,7 +7,7 @@ export const metadata = {
   title: "Sundsvallspegeln | Bio, Bar, Bistro",
   description: "Generated by create next app",
   icons: {
-    shortcut: "favicon.ico",
+    shortcut: "/favicon.ico",
   },
 };
